Refactorisation dans un package de formes
