+++ src/main/java/org/graphysica/espace2d/Espace.java
@@ -16,6 +16,7 @@
  */
 package org.graphysica.espace2d;
 
+import org.graphysica.espace2d.forme.Forme;
 import com.sun.istack.internal.NotNull;
 import javafx.scene.layout.AnchorPane;
 import org.apache.commons.math3.geometry.euclidean.twod.Vector2D;
